Remove unneeded dependency library for driver
