+++ drivers/botvac/driver.js
@@ -2,7 +2,7 @@
 
 const Homey = require('homey');
 const axios = require('axios').default;
-const BotvacUserLib = require('../../lib/BotvacUser');
+const NodeBotvacClient = require('node-botvac/lib/client');
 
 let AccessToken;
 
@@ -36,6 +36,24 @@ class BotVacDriver extends Homey.Driver {
     }
   }
 
+  async getAllRobots(accessToken) {
+    const botvacClient = new NodeBotvacClient(accessToken, 'OAuth');
+    const robotPromise = new Promise((resolve, reject) => {
+      botvacClient.getRobots((error, robots) => {
+        if (error) {
+          // eslint-disable-next-line prefer-promise-reject-errors
+          reject(`Error getting robots: ${error}`);
+        } else if (!robots) {
+          // eslint-disable-next-line prefer-promise-reject-errors
+          reject('0 robots returned');
+        } else {
+          resolve(robots);
+        }
+      });
+    });
+    return robotPromise;
+  }
+
   async listDevices() {
     if (!AccessToken) {
       throw new Error('Cannot list Neato devices!');
@@ -43,8 +61,7 @@ class BotVacDriver extends Homey.Driver {
 
     try {
       // Check if we can get the list of devices. If not throw error
-      const BotvacUser = new BotvacUserLib();
-      const robots = await BotvacUser.getAllRobots(AccessToken);
+      const robots = await this.getAllRobots(AccessToken);
       const devices = robots.map((robot) => {
         return {
           name: robot.name,
@@ -60,8 +77,8 @@ class BotVacDriver extends Homey.Driver {
         return [];
       }
       return devices;
-    } catch (err) {
-      throw new Error('Cannot list Neato devices!');
+    } catch (error) {
+      throw new Error(error);
     }
   }
 
@@ -76,7 +93,6 @@ class BotVacDriver extends Homey.Driver {
         // ... swap your code here for an access token
         const tokensObject = await this.getOauthToken(code);
         if (tokensObject && tokensObject.access_token) {
-          // BotvacUser = new BotvacUserLib(tokensObject.access_token);
           AccessToken = tokensObject.access_token;
         } else {
           throw new Error('Cannot get access token from Neato!');
